+++ api/package.json
@@ -21,7 +21,7 @@
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
     "test:e2e": "docker compose -f ../docker-compose.yml -f ../docker-compose-local.yml up --abort-on-container-exit",
     "prepare": "cd .. && [ -d '.husky' ] && husky install || true",
-    "seed": "ts-node -r tsconfig-paths/register ./node_modules/typeorm-extension/bin/cli.cjs seed:run -d ./src/database/seeds/dataSource.ts"
+    "seed": "ts-node -r tsconfig-paths/register ./node_modules/typeorm-extension/bin/cli.cjs seed:run -d src/utils/seeds/dataSource.ts"
   },
   "dependencies": {
     "@nestjs-modules/ioredis": "^2.0.1",
