+++ api/src/numerical-guidance/domain/indicator-board-meta-data.ts
@@ -1,4 +1,4 @@
-import { AggregateRoot } from 'src/building-blocks/domain/aggregate-root';
+import { AggregateRoot } from 'src/utils/building-blocks/domain/aggregate-root';
 import { IndicatorBoardMetaDataCountShouldNotExceedLimitRule } from './rule/IndicatorBoardMetaDataCountShouldNotExceedLimit.rule';
 import { IndicatorBoardMetaDataNameShouldNotEmptyRule } from './rule/IndicatorBoardMetaDataNameShouldNotEmpty.rule';
