+++ api/src/numerical-guidance/test/unit-test/domain/indicator-board-meta-data.spec.ts
@@ -1,6 +1,6 @@
 import { IndicatorBoardMetaData } from '../../../domain/indicator-board-meta-data';
 import { IndicatorBoardMetaDataCountShouldNotExceedLimitRule } from '../../../domain/rule/IndicatorBoardMetaDataCountShouldNotExceedLimit.rule';
-import { BusinessRuleValidationException } from '../../../../building-blocks/domain/business-rule-validation.exception';
+import { BusinessRuleValidationException } from '../../../../utils/building-blocks/domain/business-rule-validation.exception';
 import { IndicatorBoardMetaDataNameShouldNotEmptyRule } from '../../../domain/rule/IndicatorBoardMetaDataNameShouldNotEmpty.rule';
 
 describe('지표보드 메타데이터', () => {
