+++ api/src/utils/building-blocks/domain/aggregate-root.ts
