+++ api/src/utils/building-blocks/domain/business-rule-validation.exception.ts
