+++ api/src/utils/validation/is.interval.validation.ts
